src/requirements.txt
Flask==2.3.3
numpy==1.25.1
python-dotenv==1.0.0
scikit-learn==1.2.2
pandas==2.0.3
numpy==1.25.1
